split plot-contributions in to log parsing and plotting steps
